UpdateDid move out from lib.rs
